requirements.txt
Flask==3.0.3
opencv_python==4.9.0.80
Pillow==10.1.0
Pillow==10.3.0
Werkzeug==3.0.3
